airflow/requirements.txt
dbt-core==1.10.16
dbt-spark==1.9.3
dbt-spark[session]
dbt-spark[PyHive]
astronomer-cosmos==1.11.3

apache-airflow-providers-openlineage
